separate add property / add match
