+++ ui/src/components/query/edit/AddProperty.vue
